pyproject.toml
[build-system]
requires = [
    "setuptools",
]
build-backend = "setuptools.build_meta"

[project]
name = "extend-service-extension"
version = "0.1.0"
description = "AccelByte Extend Service Extension"
requires-python = ">=3.9"
dependencies = [
    # dependencies
    "bitarray",
    "environs",
    "Flask",
    "h2",
    "httpx",
    "mmh3",
    "requests",
    "protobuf==3.20.3",
    "PyJWT[crypto]",
    "PyYAML",
    "websockets",
    "Werkzeug",

    # accelbyte
    "accelbyte-py-sdk",

    # grpc
    "grpcio",
        # health check
        "grpcio-health-checking",
        # reflection
        "grpcio-reflection",

    # loki
    "python-logging-loki",

    # opentelemetry
    "opentelemetry-sdk",
    "opentelemetry-instrumentation-grpc",
        # prometheus
        "opentelemetry-exporter-prometheus",
        # zipkin
        "opentelemetry-exporter-zipkin-proto-http",

    # prometheus
    "prometheus-client",
]

[project.scripts]
app = "app.__main__:run"
